requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
iniconfig==2.0.0
loguru==0.6.0
packaging==23.0
pluggy==0.13.1
py==1.11.0
PyHamcrest==2.0.4
pytest==6.2.5
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dotenv==0.21.1
requests==2.28.2
selenium==3.141.0
toml==0.10.2
tqdm==4.64.1
urllib3==1.26.14
webdriver-manager==3.8.5
